backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database and ORM
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.13.1

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# WebSocket support
websockets==12.0

# Voice and AI integrations
openai==1.3.7
groq==0.4.1
deepgram-sdk==3.2.7
elevenlabs==0.2.26

# Telephony
twilio==8.10.3
telnyx==2.0.0

# SIP and WebRTC
aiortc==1.6.0
pysip==0.8.0

# Real-time communication
socketio==5.10.0
python-socketio==5.10.0

# Vector database
chromadb==0.4.18

# Redis for caching
redis==5.0.1
aioredis==2.0.1

# File processing
python-docx==1.1.0
PyPDF2==3.0.1
python-magic==0.4.27

# Web scraping
playwright==1.40.0
beautifulsoup4==4.12.2

# Monitoring and logging
structlog==23.2.0
prometheus-client==0.19.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1

# Environment management
python-dotenv==1.0.0